connect react with redux to perform operations
